requirements.txt
streamlit>=1.30.0
requests>=2.31.0
llama-cpp-python>=0.2.90
